+++ apps/web/client/src/components/ui/modal/settings/domain/custom/verified.tsx
@@ -1,4 +1,4 @@
-import { useProjectManager } from '@/components/store';
+import { useProjectManager } from '@/components/store/project';
 import { Button } from '@onlook/ui/button';
 import {
     DropdownMenu,
