+++ apps/web/client/src/components/ui/modal/settings/versions/saved-version.tsx
@@ -1,4 +1,4 @@
-import { useProjectManager } from '@/components/store';
+import { useProjectManager } from '@/components/store/project';
 import { Separator } from '@onlook/ui/separator';
 import { observer } from 'mobx-react-lite';
 import { NoSavedVersions } from './EmptyState/Saved';
